patient-management/patient-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
aiokafka==0.10.0
grpcio==1.64.0
grpcio-tools==1.64.0
